requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
pypdf==3.17.4
pymupdf==1.27.2
Pillow>=10.0.0
numpy>=1.26.0
rank-bm25==0.2.2
pydantic>=2.5.0
python-multipart==0.0.6
